requirements.txt
asgiref==3.9.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
idna==3.10
jwcrypto==1.5.6
pycparser==2.22
PyJWT==2.10.1
PyLTI1p3==2.0.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
